feat(workflow): implement CR pair filtering and scheduling plugins
