+++ src/qdash/workflow/engine/task/state_manager/__init__.py
@@ -1,0 +1,5 @@
+"""TaskStateManager for managing task state and lifecycle."""
+
+from qdash.workflow.engine.task.state_manager.manager import TaskStateManager
+
+__all__ = ["TaskStateManager"]
